fix(omnigraph): make the storage-prefix helpers importable without AWS
CI collected the new test file and died with NoRegionError before running a
single case. Importing data_tier reaches components/applications/eks ->
lib/ol_types -> lib/aws/ec2_helper, which calls boto at import time. It passed
locally only because my environment had a region configured.

The two functions are pure, so they move to a dependency-free
applications/omnigraph/storage.py — no pulumi, no boto, no ol_types — and both
data_tier and __main__ import them from there. Reproduced the failure with the
region unset and confirmed the fix the same way: 51 pass with AWS_REGION,
AWS_DEFAULT_REGION, AWS_PROFILE and both AWS config files taken away.

Also: export the resolved storage_uri, not just the prefix. The comment
claimed to give an operator "the active storage root" while exporting a knob
that is empty in the steady state and names no bucket — which is the opposite
of useful mid-migration. Both go out now: the URI for what is being served,
the prefix for what you would set or clear.
